Refactor common report stuff to BaseReport class
