+++ reports/api/daily_reservations.py
@@ -4,21 +4,55 @@ import io
 from django.utils.translation import ugettext_lazy as _
 from django.utils import formats
 from django.utils.timezone import localtime
-from rest_framework import exceptions, renderers, serializers, status, views
-from rest_framework.response import Response
-from rest_framework.settings import api_settings
+from rest_framework import exceptions, serializers
 from docx import Document
 from docx.shared import Pt
 
 from resources.models import Reservation, Resource, Unit
+from .base import BaseReport, DocxRenderer
 
 
-class DocxRenderer(renderers.BaseRenderer):
-    media_type = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document'
-    format = 'docx'
-    charset = None
-    render_style = 'binary'
+class DailyReservationsSerializer(serializers.Serializer):
+    day = serializers.DateField(required=False)
+    unit = serializers.CharField(required=False)
+    resource = serializers.CharField(required=False)
+    include_resources_without_reservations = serializers.BooleanField(required=False, default=False)
 
+    def validate_unit(self, value):
+        try:
+            unit = Unit.objects.get(id=value)
+        except Unit.DoesNotExist:
+            raise exceptions.ValidationError(
+                _('Invalid pk "{pk_value}" - object does not exist.').format(pk_value=value)
+            )
+        return unit
+
+    def validate_resource(self, value):
+        if not value:
+            return None
+
+        resource_ids = [x.strip() for x in value.split(',')]
+        resource_qs = Resource.objects.filter(id__in=resource_ids)
+        return resource_qs
+
+    def validate(self, data):
+        unit = data.get('unit')
+        resource_qs = data.get('resource')
+
+        if not (unit or resource_qs):
+            raise exceptions.ValidationError(_('Either unit or resource is required.'))
+
+        if not resource_qs:
+            resource_qs = unit.resources.all()
+        elif unit:
+            resource_qs = resource_qs.filter(unit=unit)
+        data['resource_qs'] = resource_qs
+
+        data['day'] = data.get('day', datetime.date.today())
+        return data
+
+
+class DailyReservationsDocxRenderer(DocxRenderer):
     def render(self, data, media_type=None, renderer_context=None):
         day = data['day']
         resource_qs = data['resource_qs']
@@ -91,70 +125,9 @@ class DocxRenderer(renderers.BaseRenderer):
         return output.getvalue()
 
 
-class ReportParamSerializer(serializers.Serializer):
-    day = serializers.DateField(required=False)
-    unit = serializers.CharField(required=False)
-    resource = serializers.CharField(required=False)
-    include_resources_without_reservations = serializers.BooleanField(required=False, default=False)
+class DailyReservationsReport(BaseReport):
+    serializer_class = DailyReservationsSerializer
+    renderer_classes = (DailyReservationsDocxRenderer,)
 
-    def validate_unit(self, value):
-        try:
-            unit = Unit.objects.get(id=value)
-        except Unit.DoesNotExist:
-            raise exceptions.ValidationError(
-                _('Invalid pk "{pk_value}" - object does not exist.').format(pk_value=value)
-            )
-        return unit
-
-    def validate_resource(self, value):
-        if not value:
-            return None
-
-        resource_ids = [x.strip() for x in value.split(',')]
-        resource_qs = Resource.objects.filter(id__in=resource_ids)
-        return resource_qs
-
-    def validate(self, data):
-        unit = data.get('unit')
-        resource_qs = data.get('resource')
-
-        if not (unit or resource_qs):
-            raise exceptions.ValidationError(_('Either unit or resource is required.'))
-
-        if not resource_qs:
-            resource_qs = unit.resources.all()
-        elif unit:
-            resource_qs = resource_qs.filter(unit=unit)
-        data['resource_qs'] = resource_qs
-
-        data['day'] = data.get('day', datetime.date.today())
-        return data
-
-
-class DailyReservationsReport(views.APIView):
-    renderer_classes = (DocxRenderer,)
-
-    def get(self, request, format=None):
-        serializer = ReportParamSerializer(data=request.query_params)
-        if not serializer.is_valid():
-            return Response(
-                data=serializer.errors,
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        response = Response(serializer.validated_data)
-
-        filename = '%s-%s' % (_('day-report'), serializer.validated_data['day'])
-        response['Content-Disposition'] = 'attachment; filename=%s.%s' % (filename, request.accepted_renderer.format)
-        return response
-
-    def finalize_response(self, request, response, *args, **kwargs):
-        response = super().finalize_response(request, response, *args, **kwargs)
-
-        # use the first renderer from settings to display errors
-        if response.status_code != 200:
-            first_renderer = api_settings.DEFAULT_RENDERER_CLASSES[0]()
-            response.accepted_renderer = first_renderer
-            response.accepted_media_type = first_renderer.media_type
-
-        return response
+    def get_filename(self, request, validated_data):
+        return '%s-%s.docx' % (_('day-report'), validated_data['day'])
